ci(tracing): validate OTel agnosticism in CI and document optional monitoring group
- Add pytest.importorskip("opentelemetry") to the three OTel test files
  so they skip cleanly when the monitoring group is not installed
- Extract no-op tests (noop_without_provider) into dedicated files that
  carry no OpenTelemetry import, ensuring they always run in both CI jobs
- Add back-ci-no-monitoring CI job that installs only core deps and runs
  the full test suite, validating that the app works without OTel
- Fix inline opentelemetry imports in test functions (move to top of file)
- Document the monitoring dependency group in CONTRIBUTING.md and
  server/README.md for open-source contributors
