+++ .github/workflows/CI.yml
@@ -114,6 +114,27 @@ jobs:
         working-directory: server/
         run: uv run pytest -n auto
 
+  back-ci-no-monitoring:
+    timeout-minutes: 10
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout repository
+        uses: actions/checkout@v4
+
+      - name: Install uv
+        uses: astral-sh/setup-uv@v5
+        with:
+          enable-cache: true
+          cache-dependency-glob: "server/uv.lock"
+
+      - name: Install dependencies (without monitoring group)
+        working-directory: server/
+        run: uv sync --frozen
+
+      - name: Run Unit Tests
+        working-directory: server/
+        run: uv run pytest -n auto
+
   openapi-sync-check:
     timeout-minutes: 10
     runs-on: ubuntu-latest
