+++ server/tests/shared_kernel/test_sql_base_repository_tracing.py
@@ -1,4 +1,5 @@
 import pytest
+pytest.importorskip("opentelemetry")
 from unittest.mock import MagicMock, patch
 from opentelemetry.trace import StatusCode
 from shared_kernel.adapters.secondary.sql.sql_base_repository import SQLBaseRepository
@@ -76,13 +77,3 @@ def test_commit_and_refresh_marks_span_error_on_exception(repo):
     assert status_arg == StatusCode.ERROR
 
 
-def test_operations_are_noop_without_provider(repo):
-    """commit() and commit_and_refresh() must work with the real no-op tracer."""
-    obj = MagicMock()
-    repo.commit()
-    repo._session.commit.assert_called_once()
-
-    repo._session.reset_mock()
-    repo.commit_and_refresh(obj)
-    repo._session.commit.assert_called_once()
-    repo._session.refresh.assert_called_once_with(obj)
